[Fix #88 yet again] Clean up prevention of Apple-specific GUI popup.
The previous fix involved avoiding calls that internally invoke
`sun.awt.AppContext`. This doesn't work on JDK8, which introduced additional
uses of this class. This reverts the previous workarounds from e5a25bc
and 9495ddc in favor of setting the `apple.awt.UIElement` system
property, as introduced in a096a09.