Remove setLocale call to avoid overwriting the default locale
This is a fix for the problem that DocEnv could overwrite the default
locale. The fix #210 works with most of the locales, but some locales
are still overwritten with another one (E.g., "ja_JP" =>
"ja_JP_JP_#u-ca-japanese").

This works because DocLocale#getLocale just returns Locale.getDefault()
if empty string is given as locale name and DocEnv's default locale name
is empty.