Fix debugger inspect
The first argument passed to read-debug-command is the value that
it'll return if the user types n. In other words, it's the value that
the instrumented sexp will return. I was inadvertently passing there
the value we were supposed to inspect. Inspecting worked anyway
(because that's done in parallel) but, instead of just :inspecting,
this also changed the return value of the instrumented sexp as if the
user had used :inject.