Extract more detail from compiler exceptions
This patch re-introduces the inner 'cause' of compiler exceptions.

These exceptions are not always informative, which may be why they were
hidden in 0a57231f1. For example when the compiler is complaining about
a syntax error, the exception often refers to a line number far from the
actual syntax problem.

However, the exceptions can be quite useful, for example when they
complain about an undeclared variable.

This patch re-introduces this additional detail. I'd be equally happy
with completely removing `extract-location`.  The output from the two
approaches is similar, and perhaps less code is better.