Don't reply to debug-input messages
Since this is the same subprocess that asks for debug-input, Emacs can't
tell what the received message is about until it gets parsed. Sending a
:done message in reply to debug-input makes things much more complicated
for us on the Emacs side, because we can't just use
`accept-process-output` to wait for the next input request.

Besides, this message was never used anyway.