Add tests for info middleware and small correction
Added coverage tests for the info middleware to get 100% coverage. There
are some TODO's listed in the test code since I wasn't clear on the
precise intent of certain methods.

There was a small correction made to the format-response function (the
stringified namespace wasn't merged into the response map where it
should have been, but no bug resulted since the stringified namespace
was added in subsequent steps anyway).

Changed the format-nested function slightly as well from an explicit
loop since it was difficult to test through the generated loop code. The
new code should be equivelent to the old.