Debugger: make skip-breaks depend on current code
Previously, the var `*skip-breaks*` was storing a mode (:deeper,
:before, etc.) and a sequence of coordinates. However, it did not store
any infomation indicating what code the coordinates refered to.

This would lead to strange behavior when a an :out or :here operation
would skip over code that called another instrumented function. The
debugger would treat the coordinates in *skip-breaks* as if they
referred to the code of the called function, and would seem (to the
user) to skip randomly over some or all of the function.

The fix:

1) Store *skip-breaks* as a map containing :mode, :coor, and :code,
where :code is the string of code that the :out or :here operation was
performed on.

2) Add logic to `skip-breaks?` to only skip if the to-be-skipped code
and the currently-being-debugged code are the same.