Include cider-nrepl anyway when clj is excluded
This patch allows the cider-nrepl middleware to be injected into a
project which does not depend on `org.clojure/clojure`, if and only if
the user has explicitly provided `:exclusions [... org.clojure/clojure]`
in the `project.clj` file. This allows users who know what they're doing
to take on the responsibility of providing an appropriate build of
Clojure and thus use CIDER with forks albeit with a warning.