Expand Info middleware tests
Added tests to exercise the Javadoc URL issue [Fix #308].

Added test to see if the *remote-javadocs* functionality works (same fix as
above).

Changed some of the testing targets to internal test classes/methods
instead of testing against code outside of our control when possible. As
a result, added a java source tree under the test directory and added
this to the test-clj profile.

Brings coverage back up to 100%.