Debugger: Step in
This commit implements "step in" for the debugger.

When stepping through a call to a function stored in a var, instrument
and step in to the called function if the `:in` debugger command is
sent.

If the next breakpoint is not a function call, `:in` behaves the same as
`:next`.

The instrumentation of functions that you step in to is done in a
special way. It is not simply instrumented in the normal way, because
that would cause all subsequent calls to it to trigger the debugger.
Instead, an instrumented version of the function is compiled and put
into the var's metadata, and this instrumented version is explicitly
called.

Most normal functions can be stepped in to, including both those in
source files and in jars.

Limitations:

The debugger needs to be able to find on-disk source code for a function
in order to instrument it. Functions that have been eval'ed, but not
saved, cannot be stepped into.

Only functions that can be resolved to a var can be stepped in to.
Stepping in to multimethods, protocols, or higher-order functions (eg:
`((:f m) x)`) does not work.

Opportunities:

This implements most of the ability to instrument functions without
having to send the source code over from emacs - the var name is enough.
Perhaps we could instrument functions from the namespace browser, or
from a stacktrace view (eg: instrument all the functions that are part
of my project and appear in this stacktrace).