Throw error if sending non-bencodable message (#333)
This tries to work around frozen communications related to sending
non-bencodable information through the standard Transport. If
non-bencodable information is currently sent in the reply message, it
can only be assigned to the value of the :value key, otherwise,
Transport.send will silently fail and seize up. This patch will throw an
error to the user if this type of situation is detected.