Supply CIDER version info via nREPL "describe" op (#334)
This middleware is mostly here to address bug #202
https://github.com/clojure-emacs/cider-nrepl/issues/202. The descriptor
has a :describe-fn key, providing a function which will return CIDER's
version map. This map will be merged into the :aux map returned by
nREPL's built-in "describe" op. A regular op version of this same
functionality is provided via the "cider-version" op.