Add debug-commands map
The available debugger commands, which are sent in the "input-type" key,
are now a map from key binding to command, instead of a vector of
commands.

This allows us to use key bindings other than the initial character of
the command, so we can bind "in", "inject", and "inspect" to different
keys.