Mark stackframes from the current project in stacktrace (#403)
In order to filter to just project relevant stackframes, we need a
reliable way to know which ones come from the project. The idea here
is to get all namespaces not in jars on the classpath and then also
look for a common prefix for those. In stack traces, any namespaces
that are in that set or have the common prefix we mark as part of the
project.

I looked at several alternatives but all seemed to come up short.

- I wanted to instpect the project files, but cider-nrepl's own project
file has a nonstandard name so that was no good.

- I wanted to look at the namespaces in the stacktrace and see if
these were located on disk. To do this required a good ns -> file
location function that I didn't want to recreate and make sure it
conformed to all of the edge cases of clojure's own conventions.

- And then the running process has a system.getProperty("user.dir")
which i was inspecting against the file locations of the stack
traces. The user.dir property would change when running tests and
sometimes the files would be temp files that were difficult ot match
up, and I didn't know a rhyme or reason when they were temp files and
when they were in the normal src/ location.