Include special forms in apropos
Fixes #409.

Some special forms (e.g. `if`) are included in apropos if the `search-ns` is nil or
`clojure.core`.

Included special forms are those those with documentation in `clojure.repl`,
as well as `&`, `catch`, and `finally`, the first of which uses the
documentation of `fn` and the others that of `try`.