Find relative path for cljs & boot in info [#421]
This fixes a case where in clojurescript the `:file` is stored as an
absolute path, meaning it points to a temporary file in boot.

It uses java.nio.file.Path to find all paths with the prefixes deleted,
and filters them until it finds one on the classpath. This implicitly
finds the longest one.

If it cannot find a contracted version of the file on the classpath, it
will pass through the original absolute path which points to a temporary
file.

A small optimisation here is doing this only for boot. But that does
mean that the results will be different for boot & lein as opposed to
previously where it was clj & cljs. I've not tested if there's a
performance improvement, but it will likely reduce unexpected behaviour.

| | CLJ | CLJS |
|-|-----|----------
| Boot | `:resource` |`:resource` most of the time |
| Lein | `:resource` | `:resource` unset |

Currently cljs never has `:resource`. I have no idea if this will matter
to any consumers, only future ones if at all, and it can be enabled by
disabling the check.