Require tns/session where grab-cljs-env is used
`grab-cljs-env` peeks into the message for the resolved `:session` key,
this will only have happened if the tns session middleware has been run
over the message.

TNS is consistent in how it orders middleware, so I wasn't able to
trigger this bug, but #345 seems like it could be caused by this.