Fix CLJS tests
The `:test-cljs` profile uses ClojureScript `1.7.189`, which uses `(require
...)` instead of `(:require ...)` in the ns form of `base64_vlq.clj`. This is
fixed in clojure/clojurescript@48b83d9.

b63b153 tries to correct this by specifying a higher ClojureScript version in
the `:1.9` profile, but this fails because the `:test-cljs` profile comes
later.

We resolve this by matching each version of Clojure with a compatible version
of Clojurescript in the version profiles.