[Fix #500] Use lein-sysutils to normalize java version detection
`lein-sysutils` offloads the parsing of the JDK version number to the
commons-lang3 library. The benefit is that we don't have to invent
tricky regex syntax and anticipate new versioning conventions such as
JEP 223 (http://openjdk.java.net/jeps/223).

`lein-sysutils` uses reflection to parse the available fields in
org.apache.commons.lang3.SystemUtils, which means this mechanism
should be fairly future-proof as long as the underlying library is
actively maintained.