Honor clojure.tools.namespace config and conventions
The directories to scan for a refresh can be configured with
`clojure.tools.namespace/set-refresh-dirs`. This makes CIDER honor this
configuration when set.

Specific namespaces can be excluded from being loaded or unloaded by setting
metadata on the namespace, for which there are convenience methods in
clojure.tools.namespace.repl. The docstrings of these methods say that setting
:clojure.tools.namespace.repl/load false should imply
:clojure.tools.namespace.repl/unload false. This makes CIDER honor that
convention.