Use multimethod for `cider.nrepl.middleware.test/report`
Partially address clojure-emacs/cider#2215.

Users should be able to define new message types.

With this change, such new messages will only be handled properly if `report`
produces a map adhering to the schema of that of an existing
type (e.g. `:fail`). This could be another point of extension in the future.