Implement a middleware for adding MIME data to eval responses
This changeset implements a "content-type" middleware which detects a
number of known eval results such as files, URLs and awt Images which
have meaningful multimedia representations, and annotates eval
responses with MIME data usable by a client. As with the pretty-print
middleware, content-type annotation requires the `content-type` key be
present in the initial nREPL request even if the middleware is
installed.

This changeset also implements a JDK 1.8 compatible "slurp" request
handler, usable for reading URI addressed data off of the nREPL server
with MIME content information. Together with the "content-type"
middleware, this makes it possible for a client to detect that the
last evaluation result produced a file (or some other URI addressable
object) and request that content for display.

Backwards "compatibility" to JDK 1.7 is maintained by only executing
the JDK 1.8 dependent middleware only when the required 1.8 feature
can be found on the classpath. If the required JDK 1.8 Base64 class
isn't available, an error response is returned.