Extend coercion to match cider's send format
This works around the fact that cider.el sends `[]` for `false` in the case of
tests.  To do this, it runs `seq` against lists.  This is potentially
problematic for cases where you want to search no namespaces or vars, but these
circumstances seem far less common.

It also fixes a long-standing hidden bug where the `load?` variable was always
interpreted as `true`, due to it being `[]`.

Finally, it fixes the fact that `find-ns` was used in place of `find-var` in
the coercion.