Check for cemerick/piggieback and fall back to cider/piggieback
CIDER injects cider/piggieback automatically now while some users are
still on cemerick/piggieback. The result is that this will always
return true since cider/piggieback is on the classpath.

Notably, shadow-cljs fakes a cemerick/piggieback and places the
compiler environment there for cider-nrepl. But we see the injected
cider/piggieback and so look for the cljs compiler environment there
rather than in cemerick.