[#537] A second stab at fixing the main/init problem (#539)
My fix yesterday was not properly tested - I started nREPL from the command line, but I didn't attempt to exercise it with messages. As a result, I missed the fact I was setting passing the middleware as the handler, which kept things just as broken.

This PR contains a fix for that problem, plus a new test ns cider.nrepl.main-test that exercises the start-nrepl function by passing it a variety of valid and invalid option maps. The tests go as far as checking that two :op "clone" messages can be exchanged before declaring success.