[Fix #570] Use base class of different writers
Although the `forking-writer` *should be* a `PrintWriter`, in practice we've
seen that a `BufferedWriter` can be returned. This led to the error:

>java.lang.ClassCastException: java.io.BufferedWriter cannot be cast to
java.io.PrintWriter

These types of errors are avoided by using the base class `Writer` of the
`BufferedWriter` and `PrintWriter`, just as is done in the rest of this file.