Add REPL helpers: repl-session! / close-session!
Many tests make use of `session-fixture`, which sets up an nREPL session, and
tears it down again. When working on tests this makes it tedious to quickly
evaluate test forms.

These helper functions set up the nREPL session similar to `session-fixture` but
allow you to manage the start/stop lifecycle yourself, so you can start the
session once and get to work.