Fix support for ClojureScript <1.10 on JDK>=9
ClojureScript prior to 1.10 relied on the javax.xml.bind API. Since JDK 9 this
API is part of a Jigsaw module that needs to be included when booting the JVM,
from JDK 11 this API is no longer bundled with the JDK.

To prevent users from running into this issue, we always include the JAXB API as
a separate dependency. This can be dropped once we drop support for
ClojureScript < 1.10.