Fix ClojureScript macroexpand failures
The `with-cljs-env` macro tried to grab the ClojureScript compiler env during
macroexpansion, at which point it may not be set yet. This was causing issues
where the ClojureScript compiler found an `*env*` of `nil`, causing a
NullPointerException, which was then getting swallowed by the test code.

Instead, grab the env at runtime.