Simplify the pprint logic
After working on the new pprint implementation for evaluation I
realized that the pprint picture in cider-nrepl was much more complex
and intertwined than I realized. In particular:

* there was a second pprint middleware that was simply resolving
`:pprint-fn` keys in requests to vars and updating the message map
* the same pprint logic, that was used for evaluation, was also used
for formatting edn, printing stacktraces, etc.
* there was a lot of reliance on the dynamic vars defined by
  clojure.pprint

I've opted to cut a lot of the old code and to adopt the same
approach which I implemented for evaluation. Most importantly I've
removed all of the magic with dynamic vars and replaced it with passing
a `:print-options` map to the pretty-printing functions.