[Fix #37] Move leftover code from the info middleware to orchard
This patch copies the leftover "business logic" from the `info` middleware to
the `orchard` `info` function. This will make the `info` function in `orchard`
self-contained, which is easier to test and more portable (one could use
`orchard` without the `nrepl` protocol for instance - say for a Language Server
Protocol implementation). It also moves calls to cljs-analysis to the ones in
the orchard.

Only completion is still done in cljs-tooling.