Upgrade mranderson to 0.5.0 (#604)
mranderson 0.5.x specific config added to the project file as
well. **Resolved tree** mode is used. This means that
mranderson works with a resolved dependency tree and does not create a
deeply nested directory structure, only adds prefixes to all
dependencies. Handles all depedendencies as first level ones and
processes them in a topological order.

Add project prefix in the project file for mranderson, also some
config for **unresolved tree** if that would get used in the near
future.

Breaking changes in mranderson are followed up in Makefile and README.