plugin: ensure there is a version to check to begin with
When using dependency supplying modules such as `lein-modules`,
by the time cider-nrepl does its version check it has no indication
of version.

In this case, the code throws leaving no chance to the REPL to
be spun up.

When versions are unavailable, assume the project knows what it
is doing instead and discard it from the version check altogether.