Move maint out of src
Two issues:
1. Our CI is failing our linting process with
Directory: /home/circleci/repo/src
    File                   : maint/cider/nrepl/impl/docs.clj
    has namespace          : cider.nrepl.impl.docs
    should have namespace  : maint.cider.nrepl.impl.docs
    or should be in file(s): cider/nrepl/impl/docs.cljc
                             cider/nrepl/impl/docs.clj

this arises from maint being under src. Src is on the classpath so
maint is included as well. Meaning that the file/ns didn't work out
correctly.

2. This maintenance script has been included in jars for a
while. Since its under src/ it gets put in the jar and there's no need
for this.

There were two ways to solve this: move maint out of src or nest all
the current files under some common folder so that the lib would be
src/lib and the docs would be src/maint. Seemed a bit heavy so this
was chosen.