Fix non-recognized `recur` inside `case` in the debugger   (#663)
Case expands the non-default branches into a map. This can be seen
when evaluating the following code:

`(walk/macroexpand-all '(case i
                      1 "hello"
                      "world"))`

The `contains-recur?` did so far not take that into account.

The previous `contains-recur?` also did not handle `case` when a map return
value was given correctly. Example:

(case i
   nil {:foo :bar}
   :fuss)