Fix instrumentation of small map literals. Minor doc updates.
Map literal instrumentation was relying on some heuristic before.
We use Clojures underlying array-map implementation for small map
literals and fall back to a sorted order of keys if that is not the
case. If this fails as well we simply instrument only the values of
the map in the order they appear in a sequential pass.