Use session classloader when loading deferred middleware
When loading CIDER middleware via the sideloader we need to make sure that extra
namespaces that are loaded later are also requested via the sideloader. The
original sideloader already provided a macro for
this (with-session-classloader), we just need to use it. This does mean that now
all our middlewares need to depend on the session middleware.