Document content-type ops in mw descriptor + regenerate ops.adoc
Apply the changes to the ops documentation in the middleware descriptor, then
regenerate the docs from there. Fix the location of the `:returns` map so it
shows up in the docs, and change `content-type-middleware` to `content-type` to
be in line with the others (none of them use a `-middleware` suffix in the
descriptor).